initial application: create NavController/NavView and render a simple nav view
setup structure:
* where templates go
* create Ember.application
* controller and view for the navigation

Note that we minimize what the Ember.Application.create call needs to be aware
of.  Don't want every change to have to touch this call/file.

Note also we call this._super() when we override the view's init() method.
Else, it won't render!  The init() method is invoked automatically on
Ember objects when created.