nav view with binding
Notes:
* if we try to attach the navView inside the init() for navController, then
we cannot bind the view to the controller.  The problem is we want to set
the binding using some globally accessible path, eg Example.navController.content
But, inside the init() for navController we still have not assigned to
Example.navController.  Attempting this yields an exception "Must use Ember.set to access this property"
So, we moved the attachNavView() method call to be done explictly in Example#ready()

* we created some added complexity in NavController using a closure to define
addRandomElement in order to make it accessible in a setTimeout.  Could probably
have used some jquery mechanism to setTimeout with a bound version of a instance
method on this class (binding this to the navController instance, instead of
the window).  Our point was just to show that we can add to content and
the view auto-updates (via the binding).