Add sorting and filtering
Notes:

* use a LinkView for each link in navView to simplify click handling.  Note
that when we render the LinkView objects we bind "this" (which here is the
item in the iteration) to content, so it is accessible within the view instance.
The click handler then references the object via 'content'.

* made NavController an ArrayController.  Basically, this lets us call
pushObject on the controller itself, and would allow us to use the controller
instance in a {{each}} or {{collection}} without having to reference .content

* we bind the navView to a computed property sortedLinks, which does sorting
and filtering.

* in sortedLinks, it is vital that we call toArray() on this.get('content').
I don't understand why, as the sorting seemed to work even without, but the
view would never update.  Need to investigate more.

* in sortedLinks, it's important to depend on (eg pass to .property())
"content.@each".  Otherwise, it will not refresh when items get added
to the array.  I assume this is because content itself is not changing, in
the sense that it points to the same object instance, but it is the contents
of the content array that is changing.

* in NavView, maxTextField is a class (we use .extend) but it has a lowercase
first letter.  Ember cares about case (of first letter) to determine whether
to evaluate in the global namespace (eg "Example.navController") or to
evaluate relative to the current "this" context.  So, in nav_view.hbars,
the context is the navView instance, so we reference "maxTextField" to refer
to the class defined therein.  If capitalized, we would have to provide
a full path from a global-namespace object to it, like
'Example.navController.navView.MaxTextField'